ui/organization: Feature directory for oganization outside of config-wizard
https://github.com/ethyca/fides/pull/837/files#r912233598
