+++ clients/admin-ui/src/features/config-wizard/ReviewSystemForm.tsx
@@ -18,10 +18,10 @@ import { Form, Formik } from "formik";
 import React from "react";
 
 import { useAppDispatch, useAppSelector } from "~/app/hooks";
+import { useGetOrganizationByFidesKeyQuery } from "~/features/organization";
 
 import { useGetSystemByFidesKeyQuery } from "../system/system.slice";
 import { changeReviewStep, selectSystemFidesKey } from "./config-wizard.slice";
-import { useGetOrganizationByFidesKeyQuery } from "./organization.slice";
 
 const ReviewSystemForm = ({
   handleCancelSetup,
