+++ clients/admin-ui/src/features/config-wizard/types.tsx
@@ -1,22 +1,3 @@
-export interface Organization {
-  description: string;
-  fides_key: string;
-  name: string;
-}
-
-export interface OrganizationParams {
-  page: number;
-  size: number;
-}
-
-export interface OrganizationResponse {
-  response: {};
-}
-
-export interface OrganizationUpdate {
-  response: {};
-}
-
 export type Infrastructure = "aws" | "okta";
 
 /**
