+++ clients/admin-ui/src/features/organization/index.ts
@@ -1,0 +1,2 @@
+export * from "./organization.slice";
+export * from "./types";
